Move bughunt profiling helpers into module
